Stop printing the media formats.
OpalTranscoder is not available when building with --disable-video, so
instead of placing ifdefs all over the place, the variable
allMediaFormats has been removed.  The media format information doesn't
seem that useful, so I think it can be safely removed.  If this
information turns out to be necessary, maybe we can print it out a
different way.