Remove functions and members from TkOpalManager related to placing calls on hold.
I think the code for switching from a held call to an active call was a
hold over from our original proof of concept.  The hold functionality as
previously implemented is not a feature we plan to have in the final
release, so I think it is a good to remove this code now before we start
doing more complicated thinks like adding an IVR endpoint.