Use an OpalMixerEndpoint to manage audio during a call.
The remote user and IVR can now be heard over the speaker.
This completes tasks 10 and 11.

Note: TelephonyIfc's connected member variable has been removed and
TelephonyIfc::IsConnected() now returns the value of
IsCallEstablished().

Known issues:
+Only one input is recorded at a time.  If the remote party and ivr are
both talking only one is recorded.  I think the solution for this is to
set up a second IVR to handle the recording.

+IVR does not stop playing when the user disconnects a call.

+Telekarma sometimes hangs after the user presses 'q' to quit.

This is the first step towards completing tasks 6, 9, and 27.