Implement return from IVR.
When in auto-hold mode pressing 'r' will now stop the IVR and return the
user to the call.  Automatically returning to the call when the remote
party presses a key has not been tested and might not work.

The logic for repeating the IVR message has been moved into the
TelephonyIfc class and is handled by Opal's builtin repeat and delay
handlers.  Although the delay doesn't seem to be working.

This completes task 6.

There is now a PlayWav() function in the TeleKarma class that lets
you specify the number of times to repeat a WAV and what the delay
should be between repeats.  It is recommended that this function be used to
complete task 8.