controller.cpp: disabled auto-delete.
The underlying PThread was being constructed to delete itself when done running (upon return from Main or a call to Terminate()). This behavior is convenient for the parent - no need to delete the thread - but it leaves a dangling pointer that led to thorny segfault problems in the GUI.

Under the new mode, the controller will need to be deleted (its running state should be checked first, and it should be kindly terminated if possible - avoid use of Terminate() or simply calling delete on it without first stopping it, which also invokes Terminate()).

I may further modify Controller with a run flag. An attractive alternative is to ignore turn number for quit actions (accept a quit action regardless of turn number).