Fix recording issues
We are doing recording through the OpalManager class rather than using our
own VXML script.  The VXML recording was not interacting well with the way we were muting
the microphone, so nothing was being recorded.  The OpalManger class only records
what the remote party is saying.  Pre-recorded prompts and things the telekarma user
says are not recorded, but this OK since we only record when the remote party is on
hold, so the telekarma user won't be talking anyway.